+++ api/app/controllers/spree/api/taxons_controller.rb
@@ -8,9 +8,11 @@ module Spree
           if params[:ids]
             @taxons = Spree::Taxon.accessible_by(current_ability, :read).where(id: params[:ids].split(','))
           else
-            @taxons = Spree::Taxon.accessible_by(current_ability, :read).ransack(params[:q]).result
+            @taxons = Spree::Taxon.accessible_by(current_ability, :read).order(:taxonomy_id, :lft).ransack(params[:q]).result
           end
         end
+
+        @taxons = @taxons.page(params[:page]).per(params[:per_page])
         respond_with(@taxons)
       end