+++ backend/spec/controllers/spree/admin/orders_controller_spec.rb
@@ -56,6 +56,21 @@ describe Spree::Admin::OrdersController do
         assigns[:order].created_by.should == controller.try_spree_current_user
       end
     end
+
+    # Regression test for #3684
+    context "#edit" do
+      it "does not refresh rates if the order is complete" do
+        order.stub :complete? => true
+        order.should_not_receive :refresh_shipment_rates
+        spree_get :edit, :id => order.number
+      end
+
+      it "does refresh the rates if the order is incomplete" do
+        order.stub :complete? => false
+        order.should_receive :refresh_shipment_rates
+        spree_get :edit, :id => order.number
+      end
+    end
   end
 
   context '#authorize_admin' do