+++ backend/spec/features/admin/orders/shipments_spec.rb
@@ -26,7 +26,6 @@ describe "Shipments" do
   context "moving variants between shipments", js: true do
     let!(:la) { create(:stock_location, name: "LA") }
     before(:each) do
-      create(:stock_location, name: "LA")
       visit spree.admin_path
       click_link "Orders"
       within_row(1) do
@@ -41,13 +40,13 @@ describe "Shipments" do
       targetted_select2 'LA', from: '#s2id_item_stock_location'
       click_icon :ok
       wait_for_ajax
-      page.should have_selector("#shipment_#{order.shipments.first.id}")
+      expect(page.find("#shipment_#{order.shipments.first.id}")).to be_present
 
       within_row(2) { click_icon 'resize-horizontal' }
       targetted_select2 "LA(#{order.reload.shipments.last.number})", from: '#s2id_item_stock_location'
       click_icon :ok
       wait_for_ajax
-      page.should have_selector("#shipment_#{order.reload.shipments.last.id}")
+      expect(page.find("#shipment_#{order.reload.shipments.last.id}")).to be_present
     end
   end
 end