+++ core/app/models/spree/shipment.rb
@@ -103,13 +103,11 @@ module Spree
     def refresh_rates
       return shipping_rates if shipped?
 
-      shipping_method_id = shipping_method.try(:id)
       self.shipping_rates = Stock::Estimator.new(order).shipping_rates(to_package)
 
-
-      if shipping_method_id
+      if shipping_method
         selected_rate = shipping_rates.detect { |rate|
-          rate.shipping_method_id == shipping_method_id
+          rate.shipping_method_id == shipping_method.id
         }
         self.selected_shipping_rate_id = selected_rate.id if selected_rate
       end
@@ -156,7 +154,7 @@ module Spree
     end
 
     def manifest
-      inventory_units.includes(:variant).group_by(&:variant).map do |variant, units|
+      inventory_units.joins(:variant).includes(:variant).group_by(&:variant).map do |variant, units|
         states = {}
         units.group_by(&:state).each { |state, iu| states[state] = iu.count }
         OpenStruct.new(variant: variant, quantity: units.length, states: states)