+++ core/spec/models/spree/line_item_spec.rb
@@ -20,6 +20,16 @@ describe Spree::LineItem do
       line_item.order.should_receive(:create_tax_charge!)
       line_item.destroy
     end
+
+    it "fetches deleted products" do
+      line_item.product.destroy
+      expect(line_item.reload.product).to be_a Spree::Product
+    end
+
+    it "fetches deleted variants" do
+      line_item.variant.destroy
+      expect(line_item.reload.variant).to be_a Spree::Variant
+    end
   end
 
   # Test for #3391
@@ -66,6 +76,65 @@ describe Spree::LineItem do
     before { line_item.price = 3.50 }
     it "returns a Spree::Money representing the price for one variant" do
       line_item.single_money.to_s.should == "$3.50"
+    end
+  end
+
+  context "has inventory (completed order so items were already unstocked)" do
+    let(:order) { Spree::Order.create }
+    let(:variant) { create(:variant) }
+
+    context "nothing left on stock" do
+      before do
+        variant.stock_items.update_all count_on_hand: 5, backorderable: false
+        order.contents.add(variant, 5)
+        order.create_proposed_shipments
+        order.finalize!
+      end
+
+      it "allows to decrease item quantity" do
+        line_item = order.line_items.first
+        line_item.quantity -= 1
+        line_item.target_shipment = order.shipments.first
+
+        line_item.save
+        expect(line_item).to have(0).errors_on(:quantity)
+      end
+
+      it "doesnt allow to increase item quantity" do
+        line_item = order.line_items.first
+        line_item.quantity += 2
+        line_item.target_shipment = order.shipments.first
+
+        line_item.save
+        expect(line_item).to have(1).errors_on(:quantity)
+      end
+    end
+
+    context "2 items left on stock" do
+      before do
+        variant.stock_items.update_all count_on_hand: 7, backorderable: false
+        order.contents.add(variant, 5)
+        order.create_proposed_shipments
+        order.finalize!
+      end
+
+      it "allows to increase quantity up to stock availability" do
+        line_item = order.line_items.first
+        line_item.quantity += 2
+        line_item.target_shipment = order.shipments.first
+
+        line_item.save
+        expect(line_item).to have(0).errors_on(:quantity)
+      end
+
+      it "doesnt allow to increase quantity over stock availability" do
+        line_item = order.line_items.first
+        line_item.quantity += 3
+        line_item.target_shipment = order.shipments.first
+
+        line_item.save
+        expect(line_item).to have(1).errors_on(:quantity)
+      end
     end
   end
 end