+++ core/spec/models/spree/stock/packer_spec.rb
@@ -14,6 +14,11 @@ module Spree
           packages.size.should eq 1
           packages.first.contents.size.should eq 5
         end
+
+        it 'allows users to set splitters to an empty array' do
+          packages = Packer.new(stock_location, order, []).packages
+          packages.size.should eq 1
+        end
       end
 
       context 'default_package' do
@@ -37,6 +42,22 @@ module Spree
 
           it "builds an empty package" do
             packer.default_package.contents.should be_empty
+          end
+        end
+
+        context "doesn't track inventory levels" do
+          let(:order) { Order.create }
+          let!(:line_item) { order.contents.add(create(:variant), 30) }
+
+          before { Config.track_inventory_levels = false }
+
+          it "doesn't bother stock items status in stock location" do
+            expect(subject.stock_location).not_to receive(:fill_status)
+            subject.default_package
+          end
+
+          it "still creates package with proper quantity" do
+            expect(subject.default_package.quantity).to eql 30
           end
         end
       end